Update to newer syscall

+++ nvmed/src/scheme.rs
@@ -424,38 +424,38 @@ impl SchemeBlockMut for DiskScheme {
         }
     }
 
-    fn seek(&mut self, id: usize, pos: usize, whence: usize) -> Result<Option<usize>> {
+    fn seek(&mut self, id: usize, pos: isize, whence: usize) -> Result<Option<isize>> {
         match *self.handles.get_mut(&id).ok_or(Error::new(EBADF))? {
             Handle::List(ref mut handle, ref mut size) => {
-                let len = handle.len() as usize;
+                let len = handle.len() as isize;
                 *size = match whence {
                     SEEK_SET => cmp::min(len, pos),
                     SEEK_CUR => {
-                        cmp::max(0, cmp::min(len as isize, *size as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, *size as isize + pos))
                     }
                     SEEK_END => {
-                        cmp::max(0, cmp::min(len as isize, len as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, len + pos))
                     }
                     _ => return Err(Error::new(EINVAL)),
-                };
+                } as usize;
 
-                Ok(Some(*size))
+                Ok(Some(*size as isize))
             }
             Handle::Disk(number, ref mut size) => {
                 let disk = self.disks.get_mut(&number).ok_or(Error::new(EBADF))?;
-                let len = (disk.as_ref().blocks * disk.as_ref().block_size) as usize;
+                let len = (disk.as_ref().blocks * disk.as_ref().block_size) as isize;
                 *size = match whence {
                     SEEK_SET => cmp::min(len, pos),
                     SEEK_CUR => {
-                        cmp::max(0, cmp::min(len as isize, *size as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, *size as isize + pos))
                     }
                     SEEK_END => {
-                        cmp::max(0, cmp::min(len as isize, len as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, len + pos))
                     }
                     _ => return Err(Error::new(EINVAL)),
-                };
+                } as usize;
 
-                Ok(Some(*size))
+                Ok(Some(*size as isize))
             }
             Handle::Partition(disk_num, part_num, ref mut size) => {
                 let disk = self.disks.get_mut(&disk_num).ok_or(Error::new(EBADF))?;
@@ -467,20 +467,20 @@ impl SchemeBlockMut for DiskScheme {
                     .get(part_num as usize)
                     .ok_or(Error::new(EBADF))?;
 
-                let len = (part.size * disk.as_ref().block_size) as usize;
+                let len = (part.size * disk.as_ref().block_size) as isize;
 
                 *size = match whence {
                     SEEK_SET => cmp::min(len, pos),
                     SEEK_CUR => {
-                        cmp::max(0, cmp::min(len as isize, *size as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, *size as isize + pos))
                     }
                     SEEK_END => {
-                        cmp::max(0, cmp::min(len as isize, len as isize + pos as isize)) as usize
+                        cmp::max(0, cmp::min(len, len + pos))
                     }
                     _ => return Err(Error::new(EINVAL)),
-                };
+                } as usize;
 
-                Ok(Some(*size))
+                Ok(Some(*size as isize))
             }
         }
     }
